refactor asset reconciliation tests
